Add `Density` as the basis of generating Null or Missing (#18)
It is not yet possible to script density, this just forms the basis for
doing so.

- Add `Density` as the basis of generating Null or Missing
- Add `present_value` to assure script reading gets present values
- Convert `SimpleRandomData` to use `RandomVariable`
- Rename `InnerValueGenerator`'s `gen_value` to `present_value`

This adds a `Density` generator to every `RandomVariable`. 

Every value draw generation is now the result of two random draws:
1. to determine whether the value is null, missing or present
2. to generate the actual value

Both draws are made on every generation regardless of the result of
density so that value generation is stable across present vs. absent
values.